fix : 찜하기 문제 해결
