+++ src/components/mustpost/read/readpost/PostImage.tsx
@@ -17,9 +17,7 @@ function PostImage({ img_url, link }: PostImageProps) {
         <Image
           src={img_url}
           alt="상품이미지"
-          className={`w-full h-auto rounded-lg border border-gray-2 ${
-            isLoading ? "hidden" : "block"
-          }`}
+          className={`w-full h-auto rounded-lg border border-gray-2 ${isLoading ? "hidden" : "block"}`}
           width={0}
           height={0}
           priority={true}
@@ -28,15 +26,9 @@ function PostImage({ img_url, link }: PostImageProps) {
 
         {link && (
           <a href={link} target="_blank" className="hidden md:block">
-            <div className=" cursor-pointer flex justify-center items-center bg-black absolute w-full h-full top-0 left-0 opacity-0 hover:opacity-100 hover:bg-opacity-50 rounded-[16px]">
+            <div className="cursor-pointer flex justify-center items-center bg-black absolute w-full h-full top-0 left-0 opacity-0 hover:opacity-100 hover:bg-opacity-50 rounded-[16px]">
               <p className="z-10 opacity-100 text-white">자취템 보러가기</p>
-              <Image
-                src="/img/icon-up-right.png"
-                alt="보러가기 버튼"
-                width={20}
-                height={20}
-                className="w-5 h-5"
-              />
+              <Image src="/img/icon-up-right.png" alt="보러가기 버튼" width={20} height={20} className="w-5 h-5" />
             </div>
           </a>
         )}
@@ -46,13 +38,7 @@ function PostImage({ img_url, link }: PostImageProps) {
             <div className="absolute bottom-[12px] flex w-full justify-center">
               <button className="md:hidden border border-gray-1 bg-black text-gray-1 bg-opacity-30 h-[30px] px-[90px] rounded-full flex items-center">
                 <p>자취템 보러가기</p>
-                <Image
-                  src="/img/icon-up-right.png"
-                  alt="보러가기 버튼"
-                  width={20}
-                  height={20}
-                  className="w-5 h-5"
-                />
+                <Image src="/img/icon-up-right.png" alt="보러가기 버튼" width={20} height={20} className="w-5 h-5" />
               </button>
             </div>
           </a>
